refactor: Organize auth and user utils into separate directories
