+++ src/utils/index.ts
@@ -1,26 +1,2 @@
-import { CheckAvailableEmail, FetchChangeEmail } from "./FetchChangeEmail";
-import {
-  CheckAvailableUsername,
-  FetchChangeUsername,
-} from "./FetchChangeUsername";
-import { FetchRefreshToken } from "./FetchRefreshToken";
-import { FetchUploadAvatar } from "./FetchUploadAvatar";
-import { FetchUserLogin } from "./FetchUserLogin";
-import { FetchUserLogout } from "./FetchUserLogout";
-import { FetchUserRegister } from "./FetchUserRegister";
-import { GetUserProfile } from "./GetUserProfile";
-import { GetUserSession } from "./GetUserSession";
-
-export {
-  CheckAvailableEmail,
-  CheckAvailableUsername,
-  FetchChangeEmail,
-  FetchChangeUsername,
-  FetchRefreshToken,
-  FetchUploadAvatar,
-  FetchUserLogin,
-  FetchUserLogout,
-  FetchUserRegister,
-  GetUserProfile,
-  GetUserSession,
-};
+export * from "./auth";
+export * from "./user";
